+++ src/ace_button/TimingStats.h
@@ -22,8 +22,8 @@ OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 SOFTWARE.
 */
 
-#ifndef ACE_SEGMENT_TIMING_STATS_H
-#define ACE_SEGMENT_TIMING_STATS_H
+#ifndef ACE_BUTTON_TIMING_STATS_H
+#define ACE_BUTTON_TIMING_STATS_H
 
 #include <stdint.h>
